+++ Classes/Model/Type/Car.php
@@ -47,6 +47,7 @@ final class Car extends AbstractType
         'isConsumableFor',
         'isRelatedTo',
         'isSimilarTo',
+        'isVariantOf',
         'itemCondition',
         'knownVehicleDamages',
         'logo',