+++ Classes/Model/Type/SomeProducts.php
@@ -42,6 +42,7 @@ final class SomeProducts extends AbstractType
         'isConsumableFor',
         'isRelatedTo',
         'isSimilarTo',
+        'isVariantOf',
         'itemCondition',
         'logo',
         'mainEntityOfPage',