+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Update schema definition to version 9.0 (#3)
+
 ## [1.8.0] - 2020-07-08
 
 ### Added